requirements.txt
hikari==2.1.1 # Discord API
hikari-lightbulb==2.3.5.post1 # Hikari addon that implements slash command handling
hikari-miru==4.2.0 # Hikari addon that implements component handling (i.e. modals)
psutil==7.0.0 # Gets info about the system (e.g. ram usage)